requirements.txt
streamlit>=1.28.0
clarifai>=10.11.0
clarifai-grpc>=10.0.0
python-dotenv>=1.0.0
pydub>=0.25.0
openai>=1.3.0

# Video processing dependencies
opencv-python>=4.8.0
ffmpeg-python>=0.2.0  # Primary audio extraction (high performance, reliable)
moviepy>=1.0.3        # Fallback audio extraction (compatibility issues with v2.1.2+)
numpy>=1.24.0